chocolate crud almost complete

+++ src/Components/ChocolateCard.jsx
@@ -3,6 +3,7 @@
 import React from 'react';
 import { FaBeer, FaPen, FaTrashAlt } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2'
 
 const ChocolateCard = ({ chocolate, setChocolates, chocolates, control, setControl }) => {
     const { name, _id, category, country, photo } = chocolate
@@ -10,18 +11,50 @@ const ChocolateCard = ({ chocolate, setChocolates, chocolates, control, setContr
 
     const handleDelete = (id) => {
 
-        fetch(`http://localhost:5000/delete/${id}`, {
-            method: 'DELETE'
 
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+
+            if (result.isConfirmed) {
+
+
+
+                fetch(`http://localhost:5000/delete/${id}`, {
+                    method: 'DELETE'
+
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        setControl(!control)
+                        console.log(data)
+
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+
+                      
+                       
+
+                        // const filter = chocolates.filter(cf => cf._id !== id)
+
+                    })
+
+
+
+              
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                alert('Chocolate Deleted')
 
-                // const filter = chocolates.filter(cf => cf._id !== id)
-                setControl(!control)
-                console.log(data)
-            })
 
 
     }
